chore(structure): rename files for better clarity
